requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
lxml==4.9.1
Pillow==9.3.0
PyExecJS==1.5.1
requests==2.28.1
six==1.16.0
urllib3==1.26.12
